requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
langchain>=0.2.0
langchain-anthropic>=0.1.15
langgraph>=0.1.0
langchain-mcp-adapters>=0.1.0
mcp>=1.0.0
mysql-connector-python>=8.4.0
python-dotenv>=1.0.0
